requirements.txt
PyExecJS
requests
loguru
python-dotenv
retry
openpyxl
aiohttp
opencv-python
numpy
qrcode
playwright
flask